requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.0
cfgv==3.4.0
distlib==0.3.8
Django==5.1.1
django-filter==24.3
django-resized==1.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
Markdown==3.7
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
virtualenv==20.26.3
